[api] Move backend test task mixin to a lib
Move BackendTestTasks to Backend::Tests::Tasks lib
