+++ src/api/lib/backend/backend.rb
@@ -5,3 +5,4 @@ require_dependency 'connection'
 require_dependency 'file'
 require_dependency 'api'
 require_dependency 'test'
+require_dependency 'tests/tasks'
