+++ src/api/script/start_test_backend
@@ -5,7 +5,7 @@ ENV['LC_ALL'] = 'C'
 require File.expand_path(File.join(File.dirname(__FILE__), '..', 'config', 'environment'))
 require 'builder'
 
-include TestBackendTasks
+include Backend::Tests::Tasks
 
 port_src_server = 3200
 port_rep_server = 3201
@@ -376,8 +376,8 @@ scheduler_thread = Thread.new do
   #
 
   # run scheduler for i586 and x86_64 once
-  TestBackendTasks.run_scheduler('i586')
-  TestBackendTasks.run_scheduler('x86_64')
+  Backend::Tests::Tasks.run_scheduler('i586')
+  Backend::Tests::Tasks.run_scheduler('x86_64')
 
   # Inject build job results
   inject_build_job('home:Iggy', 'TestPack', '10.2', 'i586')
@@ -395,14 +395,14 @@ scheduler_thread = Thread.new do
                           File.open("#{Rails.root}/test/fixtures/backend/binary/delete_me-1.0-1.i586.rpm").read)
 
   # run scheduler again to handle the build result
-  TestBackendTasks.run_scheduler('i586')
+  Backend::Tests::Tasks.run_scheduler('i586')
 
   # copy build result
   Backend::Connection.post('/build/HiddenProject/nada/i586/packCopy?cmd=copy&opackage=pack')
   Backend::Connection.post('/build/BaseDistro/BaseDistro_repo/i586/pack2?cmd=copy&oproject=home:Iggy&orepository=10.2&opackage=TestPack')
 
   # run scheduler again to handle the copy build event
-  TestBackendTasks.run_scheduler('i586')
+  Backend::Tests::Tasks.run_scheduler('i586')
 
   # update event database, esp. binary_release table
   BinaryRelease.transaction(requires_new: true) do
